requirements.txt
# Chatterbox TTS Dependencies for Mac M4 (Apple Silicon)
# Python 3.11 recommended

# Core ML Framework - PyTorch with MPS support
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.1.0
torchaudio>=2.1.0

# Chatterbox TTS
chatterbox-tts>=0.1.2

# Web Interface
gradio>=4.0.0

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.11.0
numpy>=1.24.0

# HuggingFace
huggingface-hub>=0.19.0
transformers>=4.35.0

# Environment & Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0

# Watermark detection (included with Chatterbox)
perth

# Google Gemini API
google-generativeai>=0.3.0

# Audio mixing
pydub>=0.25.1

# HTTP requests (for URL extraction and news)
requests>=2.31.0

